slice_provider/slice-builder/requirements.txt
flask
ruamel.yaml
PyYAML
py-json
requests
multiprocess
py-logs
